# Don't recreate iterator for each series on each timestep when evaluating a query with `timestamp()`